# openapi.json
{"openapi":"3.1.0","info":{"title":"SOVEREIGN-X Core","description":"Sovereign On-Premise Agentic AI Workbench for Confidential Industrial Work","version":"1.0.0"},"paths":{"/api/v1/health":{"get":{"tags":["System & Telemetry"],"summary":"Health Check","description":"Core system liveness, airgap isolation status, and local Ollama daemon status.","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/telemetry":{"get":{"tags":["System & Telemetry"],"summary":"Get Hardware Telemetry","description":"Retrieve real-time GPU VRAM, host system RAM, CPU load, and currently active model.","operationId":"get_hardware_telemetry_api_v1_telemetry_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/models":{"get":{"tags":["Models"],"summary":"List Models","description":"List registered local models, their capabilities, and current VRAM residency state.","operationId":"list_models_api_v1_models_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ModelHealthStatus"},"type":"array","title":"Response List Models Api V1 Models Get"}}}}}}},"/api/v1/models/swap":{"post":{"tags":["Models"],"summary":"Swap Model","description":"Force an explicit model swap and VRAM eviction of inactive models.","operationId":"swap_model_api_v1_models_swap_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwapModelRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelHealthStatus"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/models/generate":{"post":{"tags":["Models"],"summary":"Generate Text","description":"Execute non-streaming local inference with deterministic routing and VRAM safety.","operationId":"generate_text_api_v1_models_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/models/generate/stream":{"post":{"tags":["Models"],"summary":"Generate Stream","description":"Execute streaming local inference yielding Server-Sent Events (SSE).","operationId":"generate_stream_api_v1_models_generate_stream_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/workspaces/{workspace_id}/tasks":{"post":{"summary":"Create Agent Task","description":"Initialize and execute a bounded agent task in the given workspace.","operationId":"create_agent_task_api_v1_workspaces__workspace_id__tasks_post","parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"type":"string","title":"Workspace Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentTaskResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/workspaces/{workspace_id}/tasks/{task_id}":{"get":{"summary":"Get Task Details","description":"Retrieve stored status and steps of an agent task.","operationId":"get_task_details_api_v1_workspaces__workspace_id__tasks__task_id__get","parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"type":"string","title":"Workspace Id"}},{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/workspaces/{workspace_id}/tasks/{task_id}/approve":{"post":{"summary":"Approve Task Action","description":"Approve or reject a pending high-risk tool action for a waiting task.","operationId":"approve_task_action_api_v1_workspaces__workspace_id__tasks__task_id__approve_post","parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"type":"string","title":"Workspace Id"}},{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApproveActionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentTaskResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/workspaces":{"get":{"summary":"List Workspaces","description":"List all local workspaces with summary document and task counts.","operationId":"list_workspaces_api_v1_workspaces_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/WorkspaceResponse"},"type":"array","title":"Response List Workspaces Api V1 Workspaces Get"}}}}}},"post":{"summary":"Create Workspace","description":"Create a new isolated local workspace sandbox folder and SQLite record.","operationId":"create_workspace_api_v1_workspaces_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWorkspaceRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkspaceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/workspaces/{workspace_id}":{"get":{"summary":"Get Workspace","description":"Retrieve single workspace details.","operationId":"get_workspace_api_v1_workspaces__workspace_id__get","parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"type":"string","title":"Workspace Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkspaceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Workspace","description":"Wipe workspace, delete vector collection points, and remove local filesystem directory.","operationId":"delete_workspace_api_v1_workspaces__workspace_id__delete","parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"type":"string","title":"Workspace Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/workspaces/{workspace_id}/documents":{"post":{"summary":"Upload Documents","description":"Multi-part upload documents and run offline ingestion pipeline into Qdrant & SQLite.","operationId":"upload_documents_api_v1_workspaces__workspace_id__documents_post","parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"type":"string","title":"Workspace Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_documents_api_v1_workspaces__workspace_id__documents_post"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"List Documents","description":"List all ingested documents in the workspace.","operationId":"list_documents_api_v1_workspaces__workspace_id__documents_get","parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"type":"string","title":"Workspace Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DocumentSummary"},"title":"Response List Documents Api V1 Workspaces  Workspace Id  Documents Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/workspaces/{workspace_id}/documents/{document_id}":{"get":{"summary":"Get Document","description":"Retrieve document details and its extracted chunks.","operationId":"get_document_api_v1_workspaces__workspace_id__documents__document_id__get","parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"type":"string","title":"Workspace Id"}},{"name":"document_id","in":"path","required":true,"schema":{"type":"string","title":"Document Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/workspaces/{workspace_id}/query":{"post":{"summary":"Query Knowledge Vault","description":"Execute semantic similarity query directly against local Qdrant vectors.","operationId":"query_knowledge_vault_api_v1_workspaces__workspace_id__query_post","parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"type":"string","title":"Workspace Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QueryResultItem"},"title":"Response Query Knowledge Vault Api V1 Workspaces  Workspace Id  Query Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/workspaces/{workspace_id}/artifacts/{filename}":{"get":{"summary":"Download Artifact","description":"Download a generated artifact (.docx, .csv, chart) strictly from the workspace artifacts jail.","operationId":"download_artifact_api_v1_workspaces__workspace_id__artifacts__filename__get","parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"type":"string","title":"Workspace Id"}},{"name":"filename","in":"path","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audit":{"get":{"summary":"List Audit Events","description":"Retrieve historical audit log events ordered by creation timestamp descending.","operationId":"list_audit_events_api_v1_audit_get","parameters":[{"name":"workspace_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by workspace ID","title":"Workspace Id"},"description":"Filter by workspace ID"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"description":"Max events to return","default":50,"title":"Limit"},"description":"Max events to return"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AuditEventResponse"},"title":"Response List Audit Events Api V1 Audit Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audit/verify":{"post":{"summary":"Verify Audit Hash Chain","description":"Perform complete cryptographic verification of the SHA-256 continuous hash chain.","operationId":"verify_audit_hash_chain_api_v1_audit_verify_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditVerificationResponse"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AgentState":{"type":"string","enum":["IDLE","PLANNING","ACTING","TOOL_EXECUTION","OBSERVATION","WAITING_APPROVAL","COMPLETED","FAILED","TIMED_OUT","BUDGET_EXHAUSTED"],"title":"AgentState"},"AgentStepRecord":{"properties":{"step_number":{"type":"integer","title":"Step Number"},"thought":{"type":"string","title":"Thought"},"tool_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tool Name"},"tool_arguments":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Tool Arguments"},"policy_decision":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Policy Decision"},"observation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Observation"},"duration_ms":{"type":"number","title":"Duration Ms","default":0.0},"status":{"type":"string","title":"Status","default":"COMPLETED"}},"type":"object","required":["step_number","thought"],"title":"AgentStepRecord"},"AgentTaskResult":{"properties":{"task_id":{"type":"string","title":"Task Id"},"workspace_id":{"type":"string","title":"Workspace Id"},"state":{"$ref":"#/components/schemas/AgentState"},"prompt":{"type":"string","title":"Prompt"},"final_answer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Final Answer"},"steps":{"items":{"$ref":"#/components/schemas/AgentStepRecord"},"type":"array","title":"Steps"},"pending_approval":{"anyOf":[{"$ref":"#/components/schemas/PendingApproval"},{"type":"null"}]},"total_steps":{"type":"integer","title":"Total Steps","default":0},"total_duration_ms":{"type":"number","title":"Total Duration Ms","default":0.0},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["task_id","workspace_id","state","prompt"],"title":"AgentTaskResult"},"ApproveActionRequest":{"properties":{"approved":{"type":"boolean","title":"Approved","description":"True to execute, False to reject"},"tool_name":{"type":"string","title":"Tool Name"},"arguments":{"additionalProperties":true,"type":"object","title":"Arguments"}},"type":"object","required":["approved","tool_name"],"title":"ApproveActionRequest"},"AuditEventResponse":{"properties":{"id":{"type":"integer","title":"Id"},"event_uuid":{"type":"string","title":"Event Uuid"},"timestamp":{"type":"string","title":"Timestamp"},"actor":{"type":"string","title":"Actor"},"workspace_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Workspace Id"},"task_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Task Id"},"event_type":{"type":"string","title":"Event Type"},"payload_json":{"type":"string","title":"Payload Json"},"client_ip":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Ip"},"previous_hash":{"type":"string","title":"Previous Hash"},"current_hash":{"type":"string","title":"Current Hash"}},"type":"object","required":["id","event_uuid","timestamp","actor","event_type","payload_json","previous_hash","current_hash"],"title":"AuditEventResponse"},"AuditVerificationResponse":{"properties":{"is_valid":{"type":"boolean","title":"Is Valid"},"total_events":{"type":"integer","title":"Total Events"},"error_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Reason"},"last_verified_hash":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Verified Hash"}},"type":"object","required":["is_valid","total_events"],"title":"AuditVerificationResponse"},"Body_upload_documents_api_v1_workspaces__workspace_id__documents_post":{"properties":{"files":{"items":{"type":"string","contentMediaType":"application/octet-stream"},"type":"array","title":"Files"},"classification":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Classification"},"enable_ocr":{"type":"boolean","title":"Enable Ocr","default":true}},"type":"object","required":["files"],"title":"Body_upload_documents_api_v1_workspaces__workspace_id__documents_post"},"CreateTaskRequest":{"properties":{"prompt":{"type":"string","minLength":3,"title":"Prompt","description":"Task prompt for the agent"},"auto_approve_high_risk":{"type":"boolean","title":"Auto Approve High Risk","description":"Whether to automatically approve HIGH risk actions","default":false}},"type":"object","required":["prompt"],"title":"CreateTaskRequest"},"CreateWorkspaceRequest":{"properties":{"name":{"type":"string","maxLength":255,"minLength":2,"title":"Name","description":"Workspace display name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Optional description"},"classification_level":{"type":"string","title":"Classification Level","description":"Data classification: 'PUBLIC', 'INTERNAL_ENGINEERING', or 'RESTRICTED_CONFIDENTIAL'","default":"INTERNAL_ENGINEERING"}},"type":"object","required":["name"],"title":"CreateWorkspaceRequest"},"DocumentSummary":{"properties":{"id":{"type":"string","title":"Id"},"filename":{"type":"string","title":"Filename"},"mime_type":{"type":"string","title":"Mime Type"},"size_bytes":{"type":"integer","title":"Size Bytes"},"sha256_hash":{"type":"string","title":"Sha256 Hash"},"page_count":{"type":"integer","title":"Page Count"},"chunk_count":{"type":"integer","title":"Chunk Count"},"ocr_applied":{"type":"boolean","title":"Ocr Applied"},"parsing_status":{"type":"string","title":"Parsing Status"},"created_at":{"type":"string","title":"Created At"}},"type":"object","required":["id","filename","mime_type","size_bytes","sha256_hash","page_count","chunk_count","ocr_applied","parsing_status","created_at"],"title":"DocumentSummary"},"GenerationRequest":{"properties":{"prompt":{"type":"string","title":"Prompt"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model"},"system_prompt":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"System Prompt"},"images":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Images"},"temperature":{"type":"number","title":"Temperature","default":0.1},"top_p":{"type":"number","title":"Top P","default":0.9},"max_tokens":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Tokens"},"format":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Format"},"keep_alive":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Keep Alive","default":"5m"},"stream":{"type":"boolean","title":"Stream","default":false}},"type":"object","required":["prompt"],"title":"GenerationRequest","description":"Universal input payload for text generation or multimodal reasoning."},"GenerationResponse":{"properties":{"model":{"type":"string","title":"Model"},"content":{"type":"string","title":"Content"},"thinking":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thinking"},"total_duration_ms":{"type":"number","title":"Total Duration Ms","default":0.0},"prompt_tokens":{"type":"integer","title":"Prompt Tokens","default":0},"completion_tokens":{"type":"integer","title":"Completion Tokens","default":0},"done":{"type":"boolean","title":"Done","default":true},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["model","content"],"title":"GenerationResponse","description":"Universal result payload returned from local model inference."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ModelCapability":{"type":"string","enum":["TEXT","REASONING","CODE","VISION","MULTIMODAL","TOOL_CALLING","JSON_MODE"],"title":"ModelCapability"},"ModelHealthStatus":{"properties":{"model_id":{"type":"string","title":"Model Id"},"installed":{"type":"boolean","title":"Installed"},"available":{"type":"boolean","title":"Available"},"is_active":{"type":"boolean","title":"Is Active"},"capabilities":{"items":{"$ref":"#/components/schemas/ModelCapability"},"type":"array","title":"Capabilities"},"provider":{"type":"string","title":"Provider"},"last_invoked_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Invoked At"},"vram_allocated_mb":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Vram Allocated Mb"},"error_state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error State"}},"type":"object","required":["model_id","installed","available","is_active","capabilities","provider"],"title":"ModelHealthStatus","description":"Health, residency, and hardware allocation summary for a registered model."},"PendingApproval":{"properties":{"tool_name":{"type":"string","title":"Tool Name"},"arguments":{"additionalProperties":true,"type":"object","title":"Arguments"},"risk_level":{"type":"string","title":"Risk Level"},"reason":{"type":"string","title":"Reason"}},"type":"object","required":["tool_name","arguments","risk_level","reason"],"title":"PendingApproval"},"QueryRequest":{"properties":{"query":{"type":"string","minLength":1,"title":"Query","description":"Semantic search query"},"top_k":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Top K","description":"Max results to retrieve","default":4},"document_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Document Id","description":"Optional document UUID filter"}},"type":"object","required":["query"],"title":"QueryRequest"},"QueryResultItem":{"properties":{"chunk_id":{"type":"string","title":"Chunk Id"},"score":{"type":"number","title":"Score"},"content":{"type":"string","title":"Content"},"document_id":{"type":"string","title":"Document Id"},"filename":{"type":"string","title":"Filename"},"page_number":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Page Number"},"section_title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Section Title"},"classification":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Classification"}},"type":"object","required":["chunk_id","score","content","document_id","filename"],"title":"QueryResultItem"},"SwapModelRequest":{"properties":{"target_model_id":{"type":"string","title":"Target Model Id"}},"type":"object","required":["target_model_id"],"title":"SwapModelRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WorkspaceResponse":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"classification_level":{"type":"string","title":"Classification Level"},"storage_path":{"type":"string","title":"Storage Path"},"document_count":{"type":"integer","title":"Document Count","default":0},"task_count":{"type":"integer","title":"Task Count","default":0},"created_at":{"type":"string","title":"Created At"},"updated_at":{"type":"string","title":"Updated At"}},"type":"object","required":["id","name","description","classification_level","storage_path","created_at","updated_at"],"title":"WorkspaceResponse"}}}}